Crate oak_runtime

Oak Runtime implementation

Features

The oak-unsafe feature enables various debugging features, including data structure introspection functionality. This feature should only be enabled in development, as it destroys the privacy guarantees of the platform by providing easy channels for the exfiltration of private data.

Re-exports

pub use config::configure_and_run;

Modules

auth

Authentication functionality.

config

Functionality covering configuration of a Runtime instance.

permissions

time

Roughtime client implementation to provide a source of trusted time.

tls

Structs

AuxServer

Information for managing an associated server.

ChannelHalf

A reference to one half of a Channel

GrpcConfiguration

Configuration options related to gRPC pseudo-Nodes.

HttpConfiguration

Configuration options related to HTTP pseudo-Nodes.

NodeId

A unique internal identifier for a Node or pseudo-Node instance.

NodePrivilege

The downgrading (declassification + endorsement) privilege associated with a Node instance.

Runtime

Runtime structure for configuring and running a set of Oak Nodes.

RuntimeConfiguration

Configuration options that govern the behaviour of the Runtime and the Oak Application running inside it.

RuntimeProxy

A proxy object that binds together a reference to the underlying Runtime with a single NodeId.

SecureServerConfiguration

Configuration options for secure HTTP and gRPC pseudo-Nodes.

SignatureTable

Configuration options table related to Wasm module signatures.

Enums

ChannelHalfDirection

The direction of a ChannelHalf.

LabelReadStatus

Helper type to indicate whether retrieving a serialized label has succeeded or has failed with not enough capacity.

NodeReadStatus

Helper types to indicate whether a channel read operation has succeeded or has failed with not enough bytes_capacity and/or handles_capacity.

ReadStatus

Functions

construct_debug_id

Returns a unique debug_id consisting out of the provided name and NodeId.