Module oak_runtime::auth::oidc_utils

OpenID Connect utilities for exchanging authorisation codes for identity tokens, and decoding and validating identity tokens.

For reference, see: https://developers.google.com/identity/protocols/oauth2/openid-connect#exchangecode

Structs

Claims
Claims encoded into the identity token. Reference: https://developers.google.com/identity/protocols/oauth2/openid-connect#an-id-tokens-payload

ClientInfo
The OpenID Connect client ID and client secret contained in the donwloaded JSON file.

TokenError
Represents a token validation error.

Functions

exchange_code_for_token
Exchanges an authorisation code for an ID token and validates the token.

parse_client_info_json
Parses the content of the downloaded OpenID Connect client secret file.