Function oak_runtime::config::configure_and_run

pub fn configure_and_run(
    config: RuntimeConfiguration
) -> Result<Arc<Runtime>, OakError>

Configures a Runtime from the given RuntimeConfiguration and begins execution.

Returns a RuntimeProxy for an initial implicit Node, and a writeable [oak_abi::Handle] to send messages into the Runtime. Creating a new channel and passing the write [oak_abi::Handle] into the runtime will enable messages to be read back out from the RuntimeProxy.