Enum oak_runtime::LabelReadStatus

pub enum LabelReadStatus {
    Success(Vec<u8>),
    NeedsCapacity(usize),
}

Helper type to indicate whether retrieving a serialized label has succeeded or has failed with not enough capacity.

Variants

Success(Vec<u8>)

NeedsCapacity(usize)