Module oak_runtime::node

Modules

external

Enums

Configuration
A Configuration corresponds to an entry from a ApplicationConfiguration. It is the static implementation specific configuration shared between instances.

ConfigurationError
A enumeration for errors occuring when building Configuration from protobuf types.

Traits

Node
A trait implemented by every Node and pseudo-Node.

Functions

check_port
Checks if port is greater than 1023.

load_wasm
Loads a Wasm module into a Node configuration, returning an error if wasmi failed to load the module.