Trait oak_runtime::node::Node

pub trait Node: Send {
    fn run(
        self: Box<Self>, 
        runtime: RuntimeProxy, 
        handle: Handle, 
        notify_receiver: Receiver<()>
    );

    fn get_privilege(&self) -> NodePrivilege { ... }
}

Trait encapsulating execution of a Node or pseudo-Node.

Required methods

fn run(
self: Box<Self>,
runtime: RuntimeProxy,
handle: Handle,
notify_receiver: Receiver<()>
)

Execute the Node, using the provided Runtime reference and initial handle. The method should continue execution until the Node terminates.

notify_receiver receives a notification from the Runtime upon termination. This notification can be used by the Node to gracefully shut down.

Loading content...
Provided methods

fn get_privilege(&self) -> NodePrivilege

Loading content...