Struct oak_runtime::runtime::Runtime

pub struct Runtime { /* fields omitted */ }

Runtime structure for configuring and running a set of Oak Nodes.

Methods

impl Runtime

pub fn create(configuration: Configuration) -> Self

Creates a Runtime instance but does not start executing any Node.

pub fn run(self: Arc<Self>) -> Result<Handle, OakStatus>

Configures and runs the protobuf specified Application Configuration.

After starting a Runtime, calling Runtime::stop will send termination signals to Nodes and wait for them to terminate.

Returns a writeable Handle to send messages into the Runtime. To receive messages, creating a new channel and passing the write Handle into the Runtime will enable messages to be read back out.

pub fn is_terminating(&self) -> bool

Thread safe method for determining if the Runtime is terminating.

pub fn stop(&self)

Thread safe method for signaling termination to a Runtime and waiting for its Node threads to terminate.

pub fn new_channel(&self, node_id: NodeId, label: &Label) -> (Handle, Handle)

Creates a new Channel and returns a (writer handle, reader handle) pair.

pub fn wait_on_channels(
&self,
node_id: NodeId,
readers: &[Handle]
) -> Result<Vec<ChannelReadStatus>, OakStatus>

Waits on a slice of Handles, blocking until one of the following conditions:

If the Runtime is terminating this will return immediately with an ErrTerminated status for each channel.

If all readers are in an erroneous status, e.g. when all channels are orphaned, this will immediately return the channels statuses.

If any of the channels is able to read a message, the corresponding element in the returned vector will be set to Ok(ReadReady), with Ok(NotReady) signaling the channel has no message available

In particular, if there is at least one channel in good status and no messages on said channel available, Runtime::wait_on_channels will continue to block until a message is available.

pub fn channel_write(
&self,
node_id: NodeId,
reference: Handle,
msg: Message
) -> Result<(), OakStatus>

Write a message to a channel. Fails with OakStatus::ErrChannelClosed if the underlying channel has been orphaned.

pub fn channel_read(
&self,
node_id: NodeId,
reference: Handle
) -> Result<Option<Message>, OakStatus>

Thread safe. Read a message from a channel. Fails with OakStatus::ErrChannelClosed if the underlying channel is empty and has been orphaned.

pub fn channel_status(
&self,
node_id: NodeId,
reference: Handle
) -> Result<ChannelReadStatus, OakStatus>

Thread safe. This function returns:

ChannelReadStatus::ReadReady if there is at least one message in the channel.

ChannelReadStatus::Orphaned if there are no messages and there are no writers

ChannelReadStatus::NotReady if there are no messages but there are some writers

pub fn channel_try_read_message(
&self,
node_id: NodeId,
reference: Handle,
bytes_capacity: usize,
handles_capacity: usize
) -> Result<Option<ReadStatus>, OakStatus>

Thread safe. Reads a message from the channel if bytes_capacity and handles_capacity are large enough to accept the message. Fails with OakStatus::ErrChannelClosed if the underlying channel has been orphaned and is empty. If there was not enough bytes_capacity or handles_capacity, try_read_message returns Some(ReadStatus::NeedsCapacity(needed_bytes_capacity,needed_handles_capacity)). Does not guarantee that the next call will succeed after capacity adjustments as another thread may have read the original message.

pub fn channel_get_direction(
&self,
node_id: NodeId,
reference: Handle
) -> Result<HandleDirection, OakStatus>

Return the direction of a Handle. This is useful when reading Messages which contain Handle's.

pub fn channel_close(
&self,
node_id: NodeId,
reference: Handle
) -> Result<(), OakStatus>

Close a Handle, potentially orphaning the underlying [channel::Channel].

pub fn remove_node_id(&self, node_id: NodeId)

Remove a Node by NodeId from the Runtime.

pub fn node_create(
self: Arc<Self>,
node_id: NodeId,
module_name: &str,
entrypoint: &str,
label: &Label,
reader: Handle
) -> Result<(), OakStatus>

Thread safe method that attempts to create a Node within the Runtime corresponding to a given module name and entrypoint. The reader: ChannelReader is passed to the newly created Node.

The caller also specifies a Label, which is assigned to the newly created Node. See https://github.com/project-oak/oak/blob/master/docs/concepts.md#labels for more information on labels.

This method is defined on Arc and not Runtime itself, so that the Arc can clone itself and be included in a RuntimeProxy object to be given to a new Node.

pub fn node_start_instance<I>(
&self,
node_id: NodeId,
node_instance: Box<dyn Node>,
label: &Label,
initial_handles: I
) -> Result<(), OakStatus> where
I: IntoIterator<Item = Handle>,

Starts a newly created Node instance, by first initializing the necessary [NodeInfo] data structure in Runtime, allowing it to access the provided Handles, then calling Node::start on the instance, and finally storing a JoinHandle from the running instance in Runtime::node_join_handles so that it can later be terminated.

pub fn new_runtime_proxy(self: Arc<Self>) -> RuntimeProxy