Trait oak_runtime::runtime::HtmlPath

pub trait HtmlPath {
    fn html_path(&self) -> String;
}

Trait that returns the path at which the debug introspection server will show a page for a data structure.

Required methods

fn html_path(&self) -> String

Loading content...