Crate oak_runtime

Oak Runtime implementation

Features

The oak_debug feature enables various debugging features, including data structure introspection functionality. This feature should only be enabled in development, as it destroys the privacy guarantees of the platform by providing easy channels for the exfiltration of private data.

Re-exports

pub use config::application_configuration;

pub use config::configure_and_run;

pub use message::Message;

pub use runtime::NodeId;

pub use runtime::Runtime;

Modules

config

message

metrics

node

proto

runtime

Structs

Handle
A reference to a [Channel]. Each Handle has an implicit direction such that it is only possible to read or write to a Handle (exclusive or).

RuntimeConfiguration
Configuration options that govern the behaviour of the Runtime itself.