Function oak_runtime::config::configure_and_run

pub fn configure_and_run(
    app_config: ApplicationConfiguration
) -> Result<(Arc<Runtime>, Handle), OakStatus>

Configure a Runtime from the given protobuf ApplicationConfiguration and begin execution. This returns an Arc reference to the created Runtime, and a writeable Handle to send messages into the Runtime. Creating a new channel and passing the write Handle into the runtime will enable messages to be read back out from the Runtime.