Module oak_runtime::time

Roughtime client implementation to provide a source of trusted time.

Structs

RoughtimeClient
A client for requesting Roughtime from multiple servers.

RoughtimeServer
Specifies the details of a Roughtime server.

Enums

RoughtimeError
Possible errors returned by the Roughtime client.

Constants

DEFAULT_MAX_RADIUS_MICROSECONDS

DEFAULT_MIN_OVERLAPPING_INTERVALS

DEFAULT_SERVER_RETRIES

DEFAULT_TIMEOUT_SECONDS

Functions

get_default_servers
Gets the default Roughtime servers in the ecosystem.

Type Definitions

MicrosSinceEpoch
Time is given as microseconds since the UNIX epoch (00:00:00 UTC on 1 January 1970). Leap seconds are linearly smeared over a 24-hour period. That is, the smear extends from UTC noon to noon over 86,401 or 86,399 SI seconds, and all the smeared seconds are the same length.